Display active users on approval assignment card
