+++ src/components/AvatarGroup.tsx
@@ -5,6 +5,7 @@ import { cn } from '@ttab/elephant-ui/utils'
 const avatarGroupVariants = cva('flex items-center contain-layout', {
   variants: {
     size: {
+      xxs: '-space-x-3',
       xs: '-space-x-3',
       sm: '-space-x-2',
       lg: '-space-x-2',
